requirements.txt
absl-py==0.9.0
args==0.1.0
asgiref==3.2.7
asn1crypto==1.3.0
astor==0.8.0
blinker==1.4
cachetools==3.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==6.7
clint==0.5.1
Django==3.0.5
django-extensions==2.2.9
djangorestframework==3.11.0
gast==0.2.2
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
marshmallow==2.21.0
numpy==1.18.2
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
Pillow==7.0.0
pathlib2==2.3.5
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dotenv==0.12.0
pytz==2019.3
PyYAML==5.3.1
raven==6.10.0
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.14.0
sqlparse==0.3.1
tabulate==0.8.7
tensorboard==2.1.0
tensorflow-cpu==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
urllib3==1.25.8
wrapt==1.12.1
